Create main.cpp
