+++ include/fundot.h
@@ -1,3 +1,9 @@
+/**
+ * @file fundot.h
+ * 
+ * @author Jiacheng Huang
+ */
+
 #ifndef FUNDOT_H
 #define FUNDOT_H
 
@@ -7,28 +13,36 @@ namespace fundot
 {
     using std::endl;
 
+    /**
+     * An interpreter class of Fundot.
+     * This class handles input and output as well as the evaluation process of Fundot code.
+     */
     class Fundot
     {
     public:
+        /**
+         * Initialize the object with a map and insert some built-in functions into it.
+         */
         Fundot() : _obj(map<Identifier, Object>()) { _init(); }
 
+        /**
+         * Evaluate an Object as code.
+         * @param obj Object to be evaluated.
+         */
         Object eval(Object &obj);
 
+        /**
+         * Start a read-eval-print loop.
+         * @param is std::istream to be extracted.
+         * @param os std::ostream to be inserted.
+         */
         void repl(istream &is, ostream &os);
-
-        friend ostream &operator<<(ostream &os, const Fundot &fdt);
 
     private:
         Object _obj;
         void _init();
         Object _add(Object &obj);
     };
-
-    ostream &operator<<(ostream &os, const Fundot &fdt)
-    {
-        os << fdt._obj;
-        return os;
-    }
 } // namespace fundot
 
 #endif
